fix: replace the old tooltip with new one
