+++ src/Molecules/Tooltip/Tooltip.types.ts
@@ -1,8 +1,6 @@
-import { TooltipPopupProps } from '@reach/tooltip';
-
-type ReachProps = Pick<TooltipPopupProps, 'label'> & Pick<TooltipPopupProps, 'ariaLabel'>;
-
 export type Props = {
+  label: React.ReactNode;
+  ariaLabel?: string;
   className?: string;
   children?: React.ReactNode;
   /** @default bottom */
@@ -11,6 +9,7 @@ export type Props = {
   inModal?: boolean;
   /** max-width in units */
   maxWidth?: number;
-} & ReachProps;
+  mode?: 'hover' | 'click';
+};
 
 export type TooltipComponent = React.FC<Props>;
