--- src/Molecules/Tooltip/Triangle/index.ts
@@ -1,3 +1,0 @@
-import { Triangle } from './Triangle';
-
-export default Triangle;
