--- src/Molecules/TooltipNew/consts.ts
@@ -1,6 +1,0 @@
-// Values below is in px
-export const BORDER_SIZE = 1;
-export const OFFSET = 4; // Space between trigger and tip of the tooltip arrow
-export const TRIANGLE_SIZE = 10;
-export const OPEN_DELAY: number = 100;
-export const CLOSE_DELAY: number = 500;
